update - switched to modular ES6
