+++ index.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <script src="scripts/index.js" defer></script>
+  <script type="module" src="scripts/index.js" defer></script>
   <link rel="stylesheet" href="styles/styles.css">
   <title>Awesome Books</title>
 </head>
