+++ modules/loadStorage.js
@@ -1,0 +1,10 @@
+export default function getLocalStorage() {
+  let books;
+
+  if (localStorage.getItem('books') === null) {
+    books = [];
+  } else {
+    books = JSON.parse(localStorage.getItem('books'));
+  }
+  return books;
+}
